rename set iterator
